chore: create custom hook
